requirements.txt
opencv-python
pillow
keras-preprocessing
scikit-learn
tensorflow
Keras
scipy
tensorflow-datasets